+++ processor/src/test/resources/expected/com/example/arez/memoize/React4j_ParameterizedMemoizeComponent.java
@@ -2,16 +2,17 @@ package com.example.arez.memoize;
 
 import arez.Arez;
 import arez.Disposable;
+import arez.Flags;
 import arez.Observer;
 import arez.annotations.ArezComponent;
 import arez.annotations.DepType;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.InjectMode;
-import arez.annotations.Memoize;
 import arez.annotations.Observe;
 import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
+import arez.annotations.PriorityOverride;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -116,17 +117,9 @@ abstract class React4j_ParameterizedMemoizeComponent extends ParameterizedMemoiz
     }
   }
 
-  @Override
-  @Memoize(
-      priority = Priority.LOWEST,
-      name = "foo",
-      keepAlive = true,
-      reportResult = false,
-      depType = DepType.AREZ_OR_NONE,
-      observeLowerPriorityDependencies = true
-  )
-  String getIcon() {
-    return super.getIcon();
+  @PriorityOverride
+  final int fooPriority() {
+    return Flags.PRIORITY_LOWEST;
   }
 
   static final class Factory {