+++ processor/src/test/resources/expected/com/example/prop/React4j_NotObservableAsNotUpdateOnChangeProp.java
@@ -2,16 +2,17 @@ package com.example.prop;
 
 import arez.Arez;
 import arez.Disposable;
+import arez.Flags;
 import arez.Observer;
 import arez.annotations.ArezComponent;
 import arez.annotations.DepType;
 import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.InjectMode;
-import arez.annotations.Memoize;
 import arez.annotations.Observe;
 import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
+import arez.annotations.PriorityOverride;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -125,12 +126,9 @@ abstract class React4j_NotObservableAsNotUpdateOnChangeProp extends NotObservabl
     }
   }
 
-  @Override
-  @Memoize(
-      priority = Priority.LOWEST
-  )
-  int someValue() {
-    return super.someValue();
+  @PriorityOverride
+  final int someValuePriority() {
+    return Flags.PRIORITY_LOWEST;
   }
 
   static final class Factory {